Moved tests into package
